requirements.txt
streamlit>=1.28
ollama>=0.3
faiss-cpu>=1.7.0
numpy>=1.24
pypdf>=4.0
python-docx>=1.0
